Save layout: Allow space, but not empty name
